pyproject.toml
[tool.poetry]
name = "my-fastapi-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.2"  # Forcer une version spécifique

pydantic = "^2.7.4"
email-validator = "^2.2.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn my_fastapi_project.main:app --reload"
